Share the linter's error analysis with the save-time prompt
The save dialog's summary hand-rolled its own position extraction
with a non-global regex, so it took the FIRST line/col in a PyYAML
message, the context mark where the enclosing block began, and "Go to
error" jumped to line 3 for an error the banner correctly placed at
line 69. It also skipped the path sanitizing and plain-language
rewrite, showing the raw scanner dump with the host's absolute path.

Extract the pure message analysis (problem-mark extraction, path
sanitizing, indent-mismatch pinpointing, the humanized rewrites) into
yaml-error-analysis.ts, consumed by both the inline linter and
summarizeValidation, which now takes the buffer and localizer so the
dialog names the same line and shows the same wording as the banner.
Validation errors get the same sanitize plus value-type cause hint
when they live in the open buffer. The split also brings
yaml-lint-backend.ts back under the file-size cap (911 to 503 lines).
